Implement actual fix
This utilizes an observable object/state object to track the changes
in tabview pages.

By doing this and updating the observable object's index, we can use
the onChanged closure to check if we need to execute any of the
underlying code.

Thanks to Mantton and this twitter post from MaximeHeckel for helping
figure this out. The post is linked below:

https://twitter.com/maximeheckel/status/1328381063287042048?s=21

Signed-off-by: kingbri <bdashore3@gmail.com>